pyproject.toml
[project]
name = "rpg-tavern"
version = "0.1.0"
description = "LLM-powered RPG engine where players narrate actions and AI shapes the world"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
